// src/frontend/src/app/api/models/scheduled-email-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ReportFrequency } from '../models/report-frequency';
import { ScheduledEmailReceiverDto } from '../models/scheduled-email-receiver-dto';
export interface ScheduledEmailDto {
  cronExpression: string;
  frequency: ReportFrequency;
  groupId: string;
  id: string;
  name: string;
  receivers: Array<ScheduledEmailReceiverDto>;
  specificDayOfMonth: (number | string | null);
}